framework: bpl: Implement BPL UCI "set section".
To support the persistent DB feature, the BPL DB UCI functions use the
`libuci` functions to provide a long-term storage system.

Set section updates the values of the given section according to the
values of the given option according to the following rules:
- An option set to an empty string will be removed from the section.
- An option that does not exist will be added to the section.
- An option that already exists will be overridden with the new value
- Options that exist but are not mentioned the given params will remain
  unchanged.

Implemented "set section" API.

Relates to JIRA PPM-5.
Signed-off-by: Itay Elenzweig <itayx.elenzweig@intel.com>