controller: db: change selected_bands type
The current implementation of selected bands mistakenly used the
existing beerocks::eFreqType as the storage type. This is not extensible
and doesn't define the full possible configurations.
Moreover, by using the newly introduced eClientSelectedBands instead,
which is used in the BML<->controller messaging, we can simplify the set
and get operations and any future extension of supported values.

Changed in node class the storage type of the selected bands and its
description.
Changed the DB APIs to use the int8_t instead of beerocks::eFreqType.
Changed the selected bands` default value from
beerocks::eFreqType::FREQ_UNKNOWN to beerocks::PARAMETER_NOT_CONFIGURED.

PPM-5.

Signed-off-by: Adam Dov <adamx.dov@intel.com>