beerocks_cli_bml: Fix client APIs
- In client_set_client changed incoming type from INT_ARG to STRING_ARG.
Since we are using optional parameters, we need to retrieve them from
the CLI as strings, then convert them to integers.

- Added int conversion in the client_set_client LOG print.
To print the incoming values in client_set_client, need to convert them
to type int, otherwise they will be printed as Char

- Added tlvf::mac_to_string in the client_get_client LOG print.
The received mac address needs to be converted to a readable format to
be printed.

Signed-off-by: Itay Elenzweig <itayx.elenzweig@intel.com>